chore: 409 - move user action view models to specific folders
