+++ lib/widgets/cards/user_action_card.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:pass_emploi_app/presentation/user_action/user_action_view_model.dart';
 import 'package:pass_emploi_app/ui/app_colors.dart';
 import 'package:pass_emploi_app/ui/margins.dart';
 import 'package:pass_emploi_app/ui/shadows.dart';
 import 'package:pass_emploi_app/ui/text_styles.dart';
 import 'package:pass_emploi_app/widgets/tags/status_tag.dart';
-
-import '../../presentation/user_action_view_model.dart';
 
 class UserActionCard extends StatelessWidget {
   final VoidCallback onTap;
