+++ lib/widgets/tags/status_tag.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pass_emploi_app/presentation/user_action_view_model.dart';
+import 'package:pass_emploi_app/presentation/user_action/user_action_view_model.dart';
 import 'package:pass_emploi_app/ui/text_styles.dart';
 
 class StatutTag extends StatelessWidget {
