+++ test/presentation/user_action/user_action_details_view_model_test.dart
@@ -5,12 +5,12 @@ import 'package:pass_emploi_app/features/user_action/update/user_action_update_a
 import 'package:pass_emploi_app/features/user_action/update/user_action_update_state.dart';
 import 'package:pass_emploi_app/models/user_action.dart';
 import 'package:pass_emploi_app/models/user_action_creator.dart';
-import 'package:pass_emploi_app/presentation/user_action_details_view_model.dart';
+import 'package:pass_emploi_app/presentation/user_action/user_action_details_view_model.dart';
 import 'package:pass_emploi_app/redux/reducers/app_reducer.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
 import 'package:redux/redux.dart';
 
-import '../doubles/fixtures.dart';
+import '../../doubles/fixtures.dart';
 
 main() {
   group("create when action has been updated ...", () {
